requirements.txt
pycryptodome
pyyaml
websockets
